# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/article_categories:
    get:
      tags:
        - ArticleCategories
      summary: List ArticleCategories
      description: Get a paginated list of ArticleCategories with optional filtering and sorting
      operationId: listArticleCategories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleCategoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ArticleCategories
      summary: Create ArticleCategory
      description: Create a new ArticleCategory
      operationId: createArticleCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleCategoryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/article_feedbacks:
    get:
      tags:
        - ArticleFeedbacks
      summary: List ArticleFeedbacks
      description: Get a paginated list of ArticleFeedbacks with optional filtering and sorting
      operationId: listArticleFeedbacks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFeedbackList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ArticleFeedbacks
      summary: Create ArticleFeedback
      description: Create a new ArticleFeedback
      operationId: createArticleFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleFeedbackRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFeedback'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    ArticleCategory:
      type: object
      required:
        - id
        - code
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    CreateArticleCategoryRequest:
      type: object
      required:
        - code
        - name
        - metadata
      properties:
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    ArticleCategoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ArticleCategory'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    ArticleFeedback:
      type: object
      required:
        - id
        - article_id
        - helpful
        - metadata
      properties:
        id:
          type: string
          format: uuid
        article_id:
          type: string
          format: uuid
        helpful:
          type: boolean
        note:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreateArticleFeedbackRequest:
      type: object
      required:
        - article_id
        - helpful
        - metadata
      properties:
        article_id:
          type: string
          format: uuid
        helpful:
          type: boolean
        note:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    ArticleFeedbackList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ArticleFeedback'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []